XIVY-18149 Implement all default methods of Map interface on MMap
Implement all default methods of the Map interface on MMap, because MMap should use the implementation of the original Map given to it. That means all default methods on the Map interface needs to be reimplemented on the MMap. Otherwise for example ConcurrentHashMap computeIfAbsent method is never called instead put and get gets called.

 Add tests for all the new methods implemented
 Add a concurrency test for computeIfAbsent with ConcucrentHashMap as original Map